+++ android/app/src/main/java/com/woowacourse/ody/presentation/address/listener/AddressSearchListener.kt
@@ -1,5 +1,5 @@
 package com.woowacourse.ody.presentation.address.listener
 
 interface AddressSearchListener {
-    fun search()
+    fun onSearch()
 }
