+++ android/app/src/main/java/com/woowacourse/ody/presentation/creation/destination/MeetingDestinationFragment.kt
@@ -79,7 +79,7 @@ class MeetingDestinationFragment : Fragment(), AddressSearchListener {
         return geoLocationUiModel?.toGeoLocation()
     }
 
-    override fun search() = AddressSearchDialog().show(parentFragmentManager, ADDRESS_SEARCH_DIALOG_TAG)
+    override fun onSearch() = AddressSearchDialog().show(parentFragmentManager, ADDRESS_SEARCH_DIALOG_TAG)
 
     private fun showSnackBar(
         @StringRes messageId: Int,
