--- android/app/src/main/java/com/woowacourse/ody/presentation/room/CopyInviteCodeListener.kt
@@ -1,5 +1,0 @@
-package com.woowacourse.ody.presentation.room
-
-interface CopyInviteCodeListener {
-    fun copyInviteCode()
-}
