+++ android/app/src/main/java/com/woowacourse/ody/presentation/room/adapter/InviteCodeViewHolder.kt
@@ -2,7 +2,7 @@ package com.woowacourse.ody.presentation.room.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import com.woowacourse.ody.databinding.ItemInviteCodeLogBinding
-import com.woowacourse.ody.presentation.room.CopyInviteCodeListener
+import com.woowacourse.ody.presentation.room.listener.CopyInviteCodeListener
 
 class InviteCodeViewHolder(private val binding: ItemInviteCodeLogBinding) :
     RecyclerView.ViewHolder(binding.root) {
