+++ android/app/src/main/java/com/woowacourse/ody/presentation/room/listener/CopyInviteCodeListener.kt
@@ -1,0 +1,5 @@
+package com.woowacourse.ody.presentation.room.listener
+
+interface CopyInviteCodeListener {
+    fun onCopyInviteCode()
+}
