+++ android/app/src/main/java/com/woowacourse/ody/presentation/room/listener/ShareListener.kt
@@ -1,5 +1,5 @@
 package com.woowacourse.ody.presentation.room.listener
 
 interface ShareListener {
-    fun share()
+    fun onShare()
 }
