+++ android/app/src/main/res/layout/activity_meeting_room.xml
@@ -16,7 +16,7 @@
 
         <variable
             name="copyInviteCodeListener"
-            type="com.woowacourse.ody.presentation.room.CopyInviteCodeListener" />
+            type="com.woowacourse.ody.presentation.room.listener.CopyInviteCodeListener" />
 
     </data>
