+++ android/app/src/main/res/layout/toolbar_log.xml
@@ -42,7 +42,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="14dp"
-                android:onClick="@{() -> shareListener.share()}"
+                android:onClick="@{() -> shareListener.onShare()}"
                 android:src="@drawable/ic_share"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
